requirements.txt
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bitarray==1.2.1
bleach==3.1.5
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cytoolz==0.10.1
decorator==4.4.2
defusedxml==0.6.0
ecdsa==0.15
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.2
eth-bloom==1.0.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-tester==0.2.0b2
eth-typing==2.2.1
eth-utils==1.8.4
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
hexbytes==0.2.0
idna==2.9
importlib-metadata==1.6.0
ipfshttpclient==0.4.13.2
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lru-dict==1.1.6
MarkupSafe==1.1.1
mistune==0.8.4
multiaddr==0.0.9
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.6
netaddr==0.7.19
notebook==6.0.3
packaging==20.3
pandocfilters==1.4.2
parsimonious==0.8.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
py-ecc==1.7.1
py-evm==0.3.0a1
py-geth==2.3.0
py-solc-x==0.8.1
pycryptodome==3.9.7
pyethash==0.1.27
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
rlp==1.2.0
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.14.0
solc==0.0.0a0
terminado==0.8.3
testpath==0.4.4
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
trie==1.4.0
typing-extensions==3.7.4.2
urllib3==1.25.9
varint==1.0.2
wcwidth==0.1.9
web3==5.9.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.1
zipp==3.1.0
